pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/veronica_wordpress_chatbot"]

[project]
name = "veronica-wordpress-chatbot"
version = "2.0.0"
description = "AI Chatbot powered by LangGraph and WordPress API"
authors = [
    {name = "Veronica Schembri", email = "your-email@example.com"}
]
dependencies = [
    "fastapi>=0.115.14",
    "uvicorn>=0.34.3",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "langgraph-cli[inmem]>=0.1.6",
    "langsmith>=0.4.4",
    "openai>=1.93.0",
    "requests>=2.32.4",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.7",
    "slowapi>=0.1.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0"
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "types-requests>=2.32.4.20250913",
]

[project.scripts]
# Rimossi gli script console che causavano errore
# start = "uvicorn main:app --host 0.0.0.0 --port 8000"
# dev = "uvicorn main:app --host 0.0.0.0 --port 8000 --reload"
# test = "python chatbot.py"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
